Extract engine transcode normalize helper
normalize is the smallest transcoding operation still embedded in engine.py after the probe split. This moves it into engine_transcode.py while re-exporting normalize from mcp_video.engine for compatibility.

Constraint: Existing callers import normalize from mcp_video.engine.

Rejected: Move resize/convert with normalize | those are broader encoding operations with more behavior and tests.

Confidence: high

Scope-risk: narrow

Directive: Continue extracting engine operation families only when each boundary stays reviewable.

Tested: ruff check --fix mcp_video/engine.py mcp_video/engine_transcode.py

Tested: /opt/homebrew/bin/python3 normalize re-export smoke

Tested: /opt/homebrew/bin/python3 -m pytest tests/test_engine_advanced.py::TestNormalize tests/test_engine.py tests/test_e2e.py -q --tb=short
